+++ package.json
@@ -41,6 +41,7 @@
     ]
   },
   "devDependencies": {
+    "file-loader": "^6.2.0",
     "prettier": "2.4.1"
   }
 }
