+++ src/Component/MainPage/Header/Header.scss
@@ -1,0 +1,68 @@
+.Header{
+  position: absolute;
+  top: 20px;
+  left: 20px;
+  right: 20px;
+  height: 55px;
+
+  .maintitle{
+    position: absolute;
+    height: 100%;
+    top: 10%;
+    width: 5%;
+    font-size: 30px;
+    left: 100px;
+    font-family: "Fira Mono", monospace;
+    font-weight: bold;
+    cursor: pointer;
+  }
+
+  .Logo{
+    position: absolute;
+    height: 50px;
+    width: 50px;
+    left: 10px;
+    cursor: pointer;
+
+    .Logo-Img{
+      position: absolute;
+      width: 100%;
+      height: 100%;
+      cursor: pointer;
+    }
+  }
+  .search-icon{
+    position: absolute;
+    height: 100%;
+    width: 30px;
+    right: 150px;
+    cursor: pointer;
+  }
+
+  .UserProfile{
+    position: absolute;
+    background: #6e6ed3;
+    height: 100%;
+    width: 5%;
+    right:10px;
+  }
+  .btn-login{
+    position: absolute;
+    right: 50px;
+    top: 10px;
+    height: 2rem;
+    padding-left: 1rem;
+    padding-right: 1rem;
+    font-size: 1rem;
+    border-radius: 1rem;
+    border: none;
+    outline: none;
+    font-weight: bold;
+    word-break: keep-all;
+    background: rgb(52, 58, 64);
+    color: white;
+    transition: all 0.125s ease-in 0s;
+    cursor: pointer;
+  }
+
+}
