+++ src/Component/MainPage/MainPage.scss
@@ -6,50 +6,6 @@
   background-color: #f8f9fa;
   overflow: scroll;
 
-  .Header{
-    background: rgb(224, 224, 108);
-    position: absolute;
-    top: 20px;
-    left: 20px;
-    right: 20px;
-    height: 55px;
-
-    .Logo{
-      position: absolute;
-      height: 100%;
-      width: 5%;
-      left: 10px;
-
-      .Logo-Img{
-        position: absolute;
-        width: 100%;
-        height: 100%;
-      }
-    }
-    .Search{
-      position: absolute;
-      background: #deb975;
-      height: 100%;
-      width: 5%;
-      right: 100px;
-    }
-    .UserProfile{
-      position: absolute;
-      background: #6e6ed3;
-      height: 100%;
-      width: 5%;
-      right:10px;
-    }
-    .Btn-Login{
-      position: absolute;
-      background: #e3a4af;
-      height: 100%;
-      width: 5%;
-      right:10px;
-    }
-
-  }
-
 }
 
 @media(min-width: 1057px) {
