+++ src/Component/SearchPage/SearchPage.scss
@@ -3,71 +3,31 @@
   margin: 0px;
   width: 100vw;
   height: 100vh;
-  background: rgb(220,255,230);
 
-  .header{
-    background: rgb(255,255,0);
+  .searchpage-icon{
     position: absolute;
-    top: 20px;
+    color: gainsboro;
+    top: 10px;
     left: 20px;
-    right: 20px;
-    height: 55px;
-
-    .logo{
-      position: absolute;
-      height: 100%;
-      width: 5%;
-      left: 10px;
-
-      .logo-img{
-        position: absolute;
-        width: 100%;
-        height: 100%;
-      }
-    }
-    .search{
-      position: absolute;
-      background: orange;
-      height: 100%;
-      width: 5%;
-      right: 100px;
-    }
-    .goback{
-      position: absolute;
-      background: fuchsia;
-      height: 100%;
-      width: 5%;
-      right: 200px;
-    }
-    .userprofile{
-      position: absolute;
-      background: blue;
-      height: 100%;
-      width: 5%;
-      right:10px;
-    }
-    .btn-login{
-      position: absolute;
-      background: pink;
-      height: 100%;
-      width: 5%;
-      right:10px;
-    }
-
+    width: 50px;
+    height: 50px;
   }
-
   .searchbox{
-    background: #61dafb;
     position: absolute;
     top: 120px;
-    left: 20px;
-    right: 20px;
-    height: 55px;
+    left: 200px;
+    right: 200px;
+    height: 70px;
+    border: 2px solid gainsboro;
 
     .search-input{
       position: absolute;
-      height:100%;
-      width: 100%;
+      font-size: 30px;
+      right: 1px;
+      height: 65px;
+      width: 770px;
+      outline: none;
+      border: none;
     }
   }
